+++ src/Domain/Automation/Support/Livewire/Actions/WaitActionComponent.php
@@ -6,6 +6,7 @@ use Carbon\CarbonInterval;
 use Illuminate\Support\Str;
 use Illuminate\Validation\Rule;
 use Spatie\Mailcoach\Domain\Automation\Support\Livewire\AutomationActionComponent;
+use Throwable;
 
 class WaitActionComponent extends AutomationActionComponent
 {
@@ -38,6 +39,25 @@ class WaitActionComponent extends AutomationActionComponent
             'unit' => $this->unit,
             'length' => $this->length,
         ];
+    }
+
+    public function getDescriptionProperty(): string
+    {
+        if (! $this->length || ! $this->unit) {
+            return '…';
+        }
+
+        if ($this->length <= 30) {
+            return $this->length.' '.Str::plural(Str::singular($this->unit), $this->length);
+        }
+
+        try {
+            $interval = CarbonInterval::{$this->unit}($this->length);
+
+            return $interval?->cascade()->forHumans() ?? '…';
+        } catch (Throwable) {
+            return '…';
+        }
     }
 
     public function rules(): array